Introduce backpack-react-scripts
