+++ client/.eslintrc
@@ -1,0 +1,15 @@
+{
+  "extends": "skyscanner",
+  "env": {
+    "browser": true,
+    "jest": true
+  },
+  "rules": {
+    "import/no-extraneous-dependencies": [
+      "error",
+      {
+        "devDependencies": true
+      }
+    ]
+  }
+}
