--- client/.gitignore
@@ -1,12 +1,0 @@
-# See http://help.github.com/ignore-files/ for more about ignoring files.
-
-# dependencies
-node_modules
-
-# production
-build
-
-# misc
-.DS_Store
-npm-debug.log
-Thumbs.db
