--- client/config/flow/file.js.flow
@@ -1,2 +1,0 @@
-// @flow
-declare export default string;
