--- client/config/webpack.config.dev.js
@@ -1,111 +1,0 @@
-var path = require('path');
-var autoprefixer = require('autoprefixer');
-var webpack = require('webpack');
-var HtmlWebpackPlugin = require('html-webpack-plugin');
-var CaseSensitivePathsPlugin = require('case-sensitive-paths-webpack-plugin');
-var paths = require('./paths');
-
-var BACKPACK = /node_modules\/bpk-/;
-
-module.exports = {
-  devtool: 'eval',
-  entry: [
-    require.resolve('webpack-dev-server/client') + '?/',
-    require.resolve('webpack/hot/dev-server'),
-    require.resolve('./polyfills'),
-    path.join(paths.appSrc, 'index')
-  ],
-  output: {
-    // Next line is not used in dev but WebpackDevServer crashes without it:
-    path: paths.appBuild,
-    pathinfo: true,
-    filename: 'static/js/bundle.js',
-    publicPath: '/'
-  },
-  resolve: {
-    extensions: ['', '.js', '.json'],
-    alias: {
-      // This `alias` section can be safely removed after ejection.
-      // We do this because `babel-runtime` may be inside `react-scripts`,
-      // so when `babel-plugin-transform-runtime` imports it, it will not be
-      // available to the app directly. This is a temporary solution that lets
-      // us ship support for generators. However it is far from ideal, and
-      // if we don't have a good solution, we should just make `babel-runtime`
-      // a dependency in generated projects.
-      // See https://github.com/facebookincubator/create-react-app/issues/255
-      'babel-runtime/regenerator': require.resolve('babel-runtime/regenerator')
-    }
-  },
-  resolveLoader: {
-    root: paths.ownNodeModules,
-    moduleTemplates: ['*-loader']
-  },
-  module: {
-    preLoaders: [
-      {
-        test: /\.js$/,
-        loader: 'eslint',
-        include: paths.appSrc,
-      }
-    ],
-    loaders: [
-      {
-        test: /\.js$/,
-        include: [paths.appSrc, BACKPACK],
-        loader: 'babel',
-        query: require('./babel.dev')
-      },
-      {
-        test: /\.css$/,
-        include: [paths.appSrc, paths.appNodeModules],
-        loader: 'style!css!postcss'
-      },
-      {
-        test: /.*\.scss$/,
-        include: [paths.appSrc, paths.appNodeModules, BACKPACK],
-        loader: 'style!css!postcss!sass'
-      },
-      {
-        test: /\.json$/,
-        include: [paths.appSrc, paths.appNodeModules],
-        loader: 'json'
-      },
-      {
-        test: /\.(jpg|png|gif|eot|svg|ttf|woff|woff2)(\?.*)?$/,
-        exclude: [BACKPACK],
-        include: [paths.appSrc, paths.appNodeModules],
-        loader: 'file',
-        query: {
-          name: 'static/media/[name].[ext]'
-        }
-      },
-      {
-        test: /\.(mp4|webm)(\?.*)?$/,
-        include: [paths.appSrc, paths.appNodeModules],
-        loader: 'url',
-        query: {
-          limit: 10000,
-          name: 'static/media/[name].[ext]'
-        }
-      }
-    ]
-  },
-  eslint: {
-    configFile: path.join(__dirname, 'eslint.js'),
-    useEslintrc: false
-  },
-  postcss: function() {
-    return [autoprefixer];
-  },
-  plugins: [
-    new HtmlWebpackPlugin({
-      inject: true,
-      template: paths.appHtml,
-      favicon: paths.appFavicon,
-    }),
-    new webpack.DefinePlugin({ 'process.env.NODE_ENV': '"development"' }),
-    // Note: only CSS is currently hot reloaded
-    new webpack.HotModuleReplacementPlugin(),
-    new CaseSensitivePathsPlugin()
-  ]
-};
