+++ client/src/components/App/App.scss
@@ -1,0 +1,7 @@
+@import '~bpk-mixins';
+
+.App {
+  &__main {
+    padding: $bpk-spacing-sm;
+  }
+}
