+++ client/src/components/App/App.test.jsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import ShallowRenderer from 'react-test-renderer/shallow';
+
+import App from './App';
+
+describe('App', () => {
+  it('should render without crashing', () => {
+    const div = document.createElement('div');
+
+    ReactDOM.render(<App />, div);
+  });
+
+  it('should render correctly', () => {
+    const renderer = new ShallowRenderer();
+    renderer.render(<App />);
+
+    expect(renderer.getRenderOutput()).toMatchSnapshot();
+  });
+});
+
