+++ client/src/components/App/index.js
@@ -1,0 +1,3 @@
+import App from './App';
+
+export default App;
