+++ client/src/components/Header/Header.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+import logo from './logo.svg';
+import STYLES from './Header.scss';
+
+const c = className => STYLES[className] || 'UNKNOWN';
+
+const Header = () => (
+  <header className={c('Header')}>
+    <a href="/">
+      <span className={c('Header__hidden-text')}>Skyscanner</span>
+      <img className={c('Header__logo-image')} alt="Skyscanner" src={logo} />
+    </a>
+    {/* TODO burger menu icon */}
+  </header>
+);
+
+export default Header;
