--- client/src/components/topnav/TopNav.scss
@@ -1,19 +1,0 @@
-@import "~bpk-mixins";
-
-.header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: $bpk-spacing-sm $bpk-spacing-sm;
-
-  background-color: $bpk-color-white;
-  box-shadow: 0 0 10px 0 $bpk-color-gray-300;
-}
-
-.logo {
-  vertical-align: middle;
-}
-
-.logoText {
-  @include bpk-visually-hidden;
-}
