--- client/src/components/topnav/index.js
@@ -1,3 +1,0 @@
-import TopNav from './TopNav.js'
-
-export default TopNav
