+++ package.json
@@ -1,68 +1,15 @@
 {
-  "name": "app",
-  "version": "0.0.1",
   "private": true,
-  "devDependencies": {
-    "autoprefixer": "6.3.7",
-    "babel-core": "6.11.4",
-    "babel-eslint": "6.1.2",
-    "babel-loader": "6.2.4",
-    "babel-plugin-syntax-trailing-function-commas": "6.8.0",
-    "babel-plugin-transform-class-properties": "6.11.5",
-    "babel-plugin-transform-object-rest-spread": "6.8.0",
-    "babel-plugin-transform-react-constant-elements": "6.9.1",
-    "babel-plugin-transform-runtime": "6.12.0",
-    "babel-preset-es2015": "6.9.0",
-    "babel-preset-es2016": "6.11.3",
-    "babel-preset-react": "6.11.1",
-    "babel-runtime": "6.11.6",
-    "case-sensitive-paths-webpack-plugin": "1.1.2",
-    "chalk": "1.1.3",
-    "cross-spawn": "4.0.0",
-    "css-loader": "0.23.1",
-    "eslint": "3.1.1",
-    "eslint-loader": "1.4.1",
-    "eslint-plugin-flowtype": "2.4.0",
-    "eslint-plugin-import": "1.12.0",
-    "eslint-plugin-jsx-a11y": "2.0.1",
-    "eslint-plugin-react": "5.2.2",
-    "extract-text-webpack-plugin": "1.0.1",
-    "file-loader": "0.9.0",
-    "filesize": "3.3.0",
-    "fs-extra": "0.30.0",
-    "global-tunnel": "^1.2.0",
-    "gzip-size": "3.0.0",
-    "html-webpack-plugin": "2.22.0",
-    "http-proxy-agent": "^1.0.0",
-    "json-loader": "0.5.4",
-    "node-sass": "^4.5.1",
-    "opn": "4.0.2",
-    "postcss-loader": "0.9.1",
-    "promise": "7.1.1",
-    "raw-loader": "^0.5.1",
-    "rimraf": "2.5.4",
-    "sass-loader": "^4.0.0",
-    "style-loader": "0.13.1",
-    "svg-url-loader": "^2.0.2",
-    "url-loader": "0.5.7",
-    "webpack": "1.13.1",
-    "webpack-dev-server": "1.14.1"
-  },
-  "dependencies": {
-    "bpk-mixins": "^11.0.0",
-    "bpk-stylesheets": "^3.2.14",
-    "es6-promise": "^3.2.1",
-    "isomorphic-fetch": "^2.2.1",
-    "lodash": "^4.15.0",
-    "react": "^15.2.1",
-    "react-dom": "^15.2.1"
-  },
+  "name": "full-stack-recruitment-test",
+  "version": "1.0.0",
+  "description": "Skyscanner full-stack recruitment test.",
   "scripts": {
-    "client": "cd client && node ./scripts/start.js",
-    "build": "cd client && node ./scripts/build.js",
-    "server": "cd server && node src/server.js"
-  },
-  "eslintConfig": {
-    "extends": "./eslint.js"
+    "install": "(cd ./client && npm install) && (cd ./server && npm install)",
+    "client": "(cd ./client && npm start)",
+    "build": "(cd ./client && npm run build)",
+    "server": "(cd ./server && npm start)",
+    "pretest": "(cd ./client && npm run lint) && (cd ./server && npm run lint)",
+    "test": "(cd ./client && CI=true npm test) && (cd ./server && CI=true npm test)",
+    "start": "echo 'Did you mean `npm run server` or `npm run client`?'"
   }
 }
