+++ server/.eslintrc
@@ -1,0 +1,7 @@
+{
+  "extends": "skyscanner",
+  "env": {
+    "node": true,
+    "jest": true
+  }
+}
