+++ server/package.json
@@ -1,0 +1,25 @@
+{
+  "private": true,
+  "name": "server",
+  "version": "1.0.0",
+  "scripts": {
+    "start": "node ./src/server.js",
+    "test": "echo 'there are no server tests yet...'",
+    "lint:fix": "npm run lint -- --fix",
+    "lint": "eslint . --ignore-path ./../.gitignore"
+  },
+  "devDependencies": {
+    "babel-eslint": "^8.2.3",
+    "eslint": "^4.19.1",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-config-skyscanner": "^3.0.0",
+    "eslint-plugin-import": "^2.11.0",
+    "eslint-plugin-jsx-a11y": "^6.0.3",
+    "eslint-plugin-react": "^7.7.0"
+  },
+  "dependencies": {
+    "express": "^4.16.3",
+    "lodash": "^4.17.5",
+    "node-fetch": "^2.1.2"
+  }
+}
