--- server/src/api/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-  livePricing: require('./live-pricing')
-};
