+++ server/src/config.js
@@ -1,13 +1,12 @@
 // obtain API key from your Skyscanner contact
-const APIKEY = process.env.APIKEY;
+const { APIKEY } = process.env;
 
 if (!APIKEY) {
-  console.error('APIKEY environment variable missing. ' +
-    'Please re-run with `APIKEY=<key> npm run server`');
+  console.error('APIKEY environment variable missing. Please re-run with `APIKEY=<key> npm run server`');
   process.exit(1);
 }
 
 module.exports = {
   apiKey: APIKEY,
-  skyscannerApi: 'http://partners.api.skyscanner.net/'
+  skyscannerApi: 'http://partners.api.skyscanner.net/',
 };
