+++ server/src/server.js
@@ -1,13 +1,11 @@
-require('isomorphic-fetch');
-require('es6-promise').polyfill();
-
 const express = require('express');
-const app = express();
-const api = require('./api/');
 
-app.use(function(req, res, next) {
-  res.header("Access-Control-Allow-Origin", "*");
-  res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
+const app = express();
+const livePricing = require('./live-pricing');
+
+app.use((req, res, next) => {
+  res.header('Access-Control-Allow-Origin', '*');
+  res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept');
   next();
 });
 
@@ -24,17 +22,16 @@ app.get('/', (req, res) => {
   http://business.skyscanner.net/portal/en-GB/Documentation/FlightsLivePricingQuickStart
 */
 app.get('/api/search', (req, res) => {
-
-  api.livePricing.search({
+  livePricing.search({
     // TODO client to provide params
     // check in api docs what client should provide
   })
-  .then((results) => {
+    .then((results) => {
     // TODO - a better format for displaying results to the client
-    console.log('TODO: transform results for consumption by client');
-    res.json(results);
-  })
-  .catch(console.error);
+      console.log('TODO: transform results for consumption by client');
+      res.json(results);
+    })
+    .catch(console.error);
 });
 
 app.listen(4000, () => {
